rebranding namicorn => @unstoppable/resolution
