+++ README.md
@@ -1,10 +1,10 @@
-# namicorn
+# Resolution
 
 A library for interacting with blockchain domain names.
 
-You can use Namicorn in a `<script>` tag from a
-[CDN](https://unpkg.com/browse/namicorn/build/index.js), or as the `namicorn`
-package on the [npm](https://www.npmjs.com/package/namicorn) registry.
+You can use Resolution in a `<script>` tag from a
+[CDN](https://unpkg.com/browse/Resolution/build/index.js), or as the `Resolution`
+package on the [npm](https://www.npmjs.com/package/Resolution) registry.
 
 Supported domain zones:
 
@@ -15,5 +15,5 @@ Supported domain zones:
 - .luxe
 - .crypto 
 
-[API Referrence](https://unstoppabledomains.github.io/namicorn/)
-[Documentation](https://docs.unstoppabledomains.com/#tag/namicorn)
+[API Referrence](https://unstoppabledomains.github.io/Resolution/)
+[Documentation](https://docs.unstoppabledomains.com/#tag/Resolution)
