+++ docs-readme.sh
@@ -4,10 +4,10 @@ set -e
 DOCS_FOLDER=./docs
 FILE_DESTINATION="$DOCS_FOLDER/README.md"
 VERSIONS="$(find $DOCS_FOLDER -maxdepth 1 -type d -not -path '*/\.*' | basename $DOCS_FOLDER/v* | sort -r)"
-BASE_URL="https://unstoppabledomains.github.io/namicorn"
+BASE_URL="https://unstoppabledomains.github.io/-unstoppabledomains-resolution/"
 CURRENT_VERSION="v$npm_package_version"
 
-echo "# Namicorn Documentation\n"
+echo "# @unstoppabledomains/resolution Documentation\n"
 echo "[Current Version ${CURRENT_VERSION}]($BASE_URL/$CURRENT_VERSION/)\n" > $FILE_DESTINATION
 echo "## Older Versions\n" >> $FILE_DESTINATION
 for dir in $VERSIONS; do
