+++ src/utils/testHelpers.ts
@@ -1,5 +1,11 @@
+import nock from 'nock';
 import {Dictionary } from '../types';
 import {ResolutionError} from '../index';
+import mockData from '../testData/mockData.json';
+
+export const MainnetUrl = 'https://mainnet.infura.io';
+export const ZilliqaUrl = 'https://api.zilliqa.com';
+export const DefaultUrl = 'https://unstoppabledomains.com/api/v1';
 
 export function mockAsyncMethod(object: any, method: string, value) {
   if (!process.env.LIVE)
@@ -37,4 +43,29 @@ export async function expectResolutionErrorCode (
     }
   }
   expect(true).toBeFalsy();
+};
+
+export function mockAPICalls(testName: string, url = MainnetUrl) {
+  if (process.env.LIVE) {
+    return;
+  }
+  const mcdt = mockData as any;
+  const mockCall = mcdt[testName] as [any];
+
+  mockCall.forEach(({ METHOD, REQUEST, RESPONSE }) => {
+    switch (METHOD) {
+      case 'POST': {
+        nock(url)
+          // .log()
+          .post('/', JSON.stringify(REQUEST))
+          .reply(200, JSON.stringify(RESPONSE));
+      }
+      default: {
+        nock(url)
+          // .log()
+          .get(REQUEST as string)
+          .reply(200, RESPONSE);
+      }
+    }
+  });
 };
